+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0] - 2022-xx-xx.
+
+* added function isValidJamo
+* added function isValidJamoCode
+
 ## [0.1.1] - 2021-05-09.
 
 * added example